requirements.txt
Flask==2.0.2
st-annotated-text
bokeh
streamlit-bokeh-events